requirements.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
blis==1.3.0
boto3==1.39.11
botocore==1.39.11
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.7.14
charset-normalizer==3.4.2
chromadb==1.0.15
ci-info==0.3.0
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
cymem==2.0.11
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
elastic-transport==8.17.1
elasticsearch==9.0.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
etelemetry==0.3.1
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keybert==0.9.0
kubernetes==33.1.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.71
langchain-google-genai==2.1.8
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.4.8
language_data==1.3.0
looseversion==1.3.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.6.3
murmurhash==1.0.13
mypy_extensions==1.1.0
neo4j==5.28.1
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==2.3.1
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.97.1
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.0
overrides==7.7.0
packaging==24.2
pandas==2.3.1
pathlib==1.0.1
pillow==11.3.0
pinecone==7.3.0
pinecone-client==6.0.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
posthog==5.4.0
preshed==3.0.10
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==5.8.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
starlette==0.47.2
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
traits==7.0.2
transformers==4.53.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
